contributors/adobe/incoming/css3-regions/interactivity/keyboard/regions-keyboard-events-007.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>CSS Regions: onkeydown and onkeyup events inside multiple regions using Tab key in forms</title>
        <link rel="author" title="David Alcala" href="mailto:dalcala@adobe.com">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#the-flow-into-property">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#the-flow-from-property">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#relation-to-document-events">
        <meta name="assert" content="CSS regions module does not alter the normal processing of events in the document
         tree. Therefore, onkeydown and onkeyup events inside multiple named flows should function correctly.">
        <meta name="flags" content="dom interact">
        <link href="support/css/regions-keyboard.css" rel="stylesheet" type="text/css">
        <link href="/resources/testharness.css" rel="stylesheet" type="text/css">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="support/js/regions-keyboard-events-test-helper.js"></script>
    </head>
    <body>
        <p id="msg">Press the Tab key four times.</p>
        <div id="content">
            <form id="my-form">
            <div class="block">
                <input type="text" id="textfield-inside1" size="10">
            </div>
            <div class="block">
                <input type="text" id="textfield-inside2" size="10">
            </div>
            <div id="target-block" class="block">
                <input type="button" id="button-inside" value="Button">
            </div>
             </form>
        </div>
        <div class="region"></div>
        <div class="region"></div>
        <div class="region"></div>
        <br/>
        <div id="outside-region">This div is outside the region
            <form id="form-outside">
                <input type="text" id="textfield-outside" size="10">
                <input type="button" id="button-outside" value="Button">
            </form>
        </div>
        <div id="log"></div>
    </body>
    <script type="text/javascript">

        if(getLeftPosition("target-block") !== 0) {

            var test1 = async_test("onkeyup event fired in textfield1 inside region", {timeout: testTimeout});
            test1.step(function () {
                var testTarget = document.getElementById("textfield-inside1");
                testTarget.onkeyup = test1.step_func(function (evt) {
                    /* Log assertion that the event got fired */
                    assert_true(true);

                    test1.done();
                });
            });

            var test2 = async_test("onkeydown event fired in textfield1 inside region", {timeout: testTimeout});
            test2.step(function () {
                var testTarget = document.getElementById("textfield-inside1");
                testTarget.onkeydown = test2.step_func(function (evt) {

                    /* Verify the target outside the region has the correct ID */
                    test(function() {
                        assert_equals(evt.target.id, "textfield-inside1");
                    }, "Textfield1 inside region has the correct ID");

                    /* Verify the textfield received focus from tab key */
                    test(function() {
                        assert_equals(document.activeElement.id, "textfield-inside1");
                    }, "Textfield1 inside region received focus");

                    test2.done();
                });
            });

            var test3 = async_test("onkeyup event fired in textfield2 inside region", {timeout: testTimeout});
            test3.step(function () {
                var testTarget = document.getElementById("textfield-inside2");
                testTarget.onkeyup = test3.step_func(function (evt) {
                    /* Log assertion that the event got fired */
                    assert_true(true);

                    test3.done();
                });
            });

            var test4 = async_test("onkeydown event fired in textfield2 inside region", {timeout: testTimeout});
            test4.step( function() {
                var testTarget = document.getElementById("textfield-inside2");
                testTarget.onkeydown = test4.step_func(function (evt) {

                    /* Verify the target outside the region has the correct ID */
                    test(function () {
                        assert_equals(evt.target.id, "textfield-inside2");
                    }, "Textfield2 inside region has the correct ID");

                    /* Verify the textfield received focus from tab key */
                    test(function () {
                        assert_equals(document.activeElement.id, "textfield-inside2");
                    }, "Textfield2 inside region received focus");

                    test4.done();
                });
            });

            var test5 = async_test("onkeyup event fired in button inside region", {timeout: testTimeout});
            test5.step( function() {
                var testTarget = document.getElementById("button-inside");
                testTarget.onkeyup = test5.step_func(function (evt) {

                    /* Verify the button received focus from tab key */
                    test(function () {
                        assert_equals(document.activeElement.id, "button-inside");
                    }, "Button inside region received focus");

                    test5.done();
                });
            });

            var test6 = async_test("onkeydown event fired in button inside region", {timeout: testTimeout});
            test6.step( function() {
                var testTarget = document.getElementById("button-inside");
                testTarget.onkeydown = test6.step_func(function (evt) {

                    /* Verify the button inside the region has the correct ID */
                    test(function () {
                        assert_equals(evt.target.id, "button-inside");
                    }, "Button inside region has the correct ID");

                    test6.done();
                });
            });

            tabKeyPresses(4);
            completionCallback();

        } //end of getLeftPosition test
        else {
            test(function () {
                assert_true(false);
            }, "Regions are not enabled");
        }
    </script>
</html>
